requirements.txt
ultralytics>=8.0.0
torch>=1.10.0
torchvision>=0.11.0
streamlit>=1.28.0
opencv-python>=4.5.0
numpy>=1.19.0
pillow>=10.0.0
easyocr>=1.7.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
pytesseract>=0.3.0
pandas>=2.0.0